requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyQt5_sip == 12.8.1
PyYAML == 5.4.1
pyperclip == 1.8.2
setuptools == 49.2.1
